Upstream version 2.5.7

Conflicts:
	.gitignore
	Makefile.in
	config.h.in
	configure
	configure.ac
	depot/Makefile.in
	doc/changes.html
	doc/install-http.html
	doc/install-upgrade.html
	doc/man/arbitron.8.html
	doc/man/chk_cyrus.8.html
	doc/man/ctl_cyrusdb.8.html
	doc/man/ctl_deliver.8.html
	doc/man/ctl_mboxlist.8.html
	doc/man/cvt_cyrusdb.8.html
	doc/man/cyr_dbtool.8.html
	doc/man/cyr_df.8.html
	doc/man/cyr_expire.8.html
	doc/man/cyr_synclog.8.html
	doc/man/cyrus.conf.5.html
	doc/man/deliver.8.html
	doc/man/fetchnews.8.html
	doc/man/fud.8.html
	doc/man/httpd.8.html
	doc/man/idled.8.html
	doc/man/imapd.8.html
	doc/man/imapd.conf.5.html
	doc/man/imclient.3.html
	doc/man/imtest.1.html
	doc/man/installsieve.1.html
	doc/man/ipurge.8.html
	doc/man/krb.equiv.5.html
	doc/man/lmtpd.8.html
	doc/man/lmtptest.1.html
	doc/man/master.8.html
	doc/man/mbexamine.8.html
	doc/man/mbpath.8.html
	doc/man/mupdatetest.1.html
	doc/man/nntpd.8.html
	doc/man/nntptest.1.html
	doc/man/notifyd.8.html
	doc/man/pop3d.8.html
	doc/man/pop3test.1.html
	doc/man/quota.8.html
	doc/man/reconstruct.8.html
	doc/man/rmnews.8.html
	doc/man/sivtest.1.html
	doc/man/smmapd.8.html
	doc/man/smtptest.1.html
	doc/man/squatter.8.html
	doc/man/sync_client.8.html
	doc/man/sync_reset.8.html
	doc/man/sync_server.8.html
	doc/man/syncnews.8.html
	doc/man/timsieved.8.html
	doc/man/tls_prune.8.html
	doc/man/unexpunge.8.html
	doc/specs.html
	doc/text/changes
	doc/text/install
	doc/text/install-http
	doc/text/install-upgrade
	doc/text/specs
	imap/Makefile.in
	imap/annotate.c
	imap/backend.c
	imap/caldav_db.c
	imap/caldav_db.h
	imap/carddav_db.c
	imap/carddav_db.h
	imap/ctl_zoneinfo.c
	imap/dav_db.c
	imap/dav_db.h
	imap/dav_reconstruct.c
	imap/dlist.c
	imap/global.c
	imap/global.h
	imap/http_caldav.c
	imap/http_carddav.c
	imap/http_dav.c
	imap/http_dav.h
	imap/http_ischedule.c
	imap/http_proxy.c
	imap/http_proxy.h
	imap/http_rss.c
	imap/http_timezone.c
	imap/httpd.c
	imap/httpd.h
	imap/idle.c
	imap/imap_proxy.c
	imap/imap_proxy.h
	imap/imapd.c
	imap/index.c
	imap/jcal.c
	imap/lmtpd.c
	imap/mailbox.c
	imap/mailbox.h
	imap/mbdump.c
	imap/mboxlist.c
	imap/mboxlist.h
	imap/mboxname.c
	imap/nntpd.c
	imap/protocol.h
	imap/reconstruct.c
	imap/sync_client.c
	imap/sync_reset.c
	imap/sync_server.c
	imap/sync_support.c
	imap/sync_support.h
	imap/tls.c
	imap/upgrade_index.c
	imap/user.c
	imap/xcal.c
	imap/zoneinfo_db.c
	lib/charset.c
	lib/cyrusdb_flat.c
	lib/imapoptions
	lib/imapopts.c
	lib/imapopts.h
	lib/prot.c
	lib/util.c
	lib/util.h
	man/Makefile.in
	man/imapd.conf.5
	man/sieveshell.1
	man/sync_client.8
	notifyd/Makefile.in
	perl/sieve/lib/lex.c
	perl/sieve/lib/lex.h
	ptclient/Makefile.in
	timsieved/Makefile.in
	tools/vzic/Makefile
	tools/vzic/vzic-output.c
	tools/vzic/vzic-output.h
	tools/vzic/vzic-parse.c
	tools/vzic/vzic.c
	xversion.h

+++ lib/map.h
@@ -38,8 +38,6 @@
  * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN
  * AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING
  * OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
- *
- * $Id: map.h,v 1.14 2010/01/06 17:01:46 murch Exp $
  */
 
 #ifndef INCLUDED_MAP_H
@@ -63,12 +61,12 @@ extern const char *map_method_desc;
  * 	of the applicable mailbox (if any), and may be NULL
  */
 extern void map_refresh(int fd, int onceonly, const char **base,
-			unsigned long *len, unsigned long newlen,
+			size_t *len, size_t newlen,
 			const char *name, const char *mboxname);
 
 /* map_free will free a memory map allocated by map_refresh
  *
  * base and len are the same values that were passed to map_refresh */
-extern void map_free(const char **base, unsigned long *len);
+extern void map_free(const char **base, size_t *len);
 
 #endif /* INCLUDED_MAP_H */
